ML/Flask-Server/requirements.txt
Flask==2.1.1
tensorflow==2.10.0
numpy==1.21.6
Pillow==9.0.1
